feat: add mvc architecture
